Fix lint issues.

+++ contracts/base/action/CompoundMintPostAction.sol
@@ -18,10 +18,10 @@ contract CompoundMintPostAction is PostActionBase {
     /**
         @notice It is used as a success result after minting the tokens.
      */
-    uint internal constant SUCCESS_CODE = 0;
+    uint256 internal constant SUCCESS_CODE = 0;
     bytes32 internal constant COMPOUND_ACTION_DATA = "CompoundAction";
     address internal constant ADDRESS_EMPTY = address(0x0);
-    
+
     /** Properties */
     ICompoundSettings public compoundSettings;
 
@@ -37,7 +37,10 @@ contract CompoundMintPostAction is PostActionBase {
         @param aCompoundSettings contract address which contains the platform configuration for the Compound.finance platform.
         @dev The Eternal Storage implementation must implement the IStorage interface.
      */
-    constructor(address storageAddress, address aCompoundSettings) public PostActionBase(storageAddress) {
+    constructor(address storageAddress, address aCompoundSettings)
+        public
+        PostActionBase(storageAddress)
+    {
         compoundSettings = ICompoundSettings(aCompoundSettings);
     }
 
@@ -64,24 +67,52 @@ contract CompoundMintPostAction is PostActionBase {
             "Balance of ERC20 is not gte amount to transfer."
         );
 
-        if(compoundSettings.supportErc20(postActionData.targetToken)) {
-            address cTargetTokenAddress = compoundSettings.getCEr20(postActionData.targetToken);
-            cTargetTokenAddress.requireNotEmpty("Target token is not supported as CErc20.");
-            
+        if (compoundSettings.supportErc20(postActionData.targetToken)) {
+            address cTargetTokenAddress = compoundSettings.getCEr20(
+                postActionData.targetToken
+            );
+            cTargetTokenAddress.requireNotEmpty(
+                "Target token is not supported as CErc20."
+            );
+
             CErc20 cTargetToken = CErc20(cTargetTokenAddress);
 
-            uint postActionInitialBalance = cTargetToken.balanceOf(address(this));
+            uint256 postActionInitialBalance = cTargetToken.balanceOf(
+                address(this)
+            );
 
-            require(targetToken.approve(cTargetTokenAddress, postActionData.toAmount), "Compound: Approve target token failed.");
+            require(
+                targetToken.approve(
+                    cTargetTokenAddress,
+                    postActionData.toAmount
+                ),
+                "Compound: Approve target token failed."
+            );
 
-            uint cTargetTokenMintResult = cTargetToken.mint(postActionData.toAmount);
-            require(cTargetTokenMintResult == SUCCESS_CODE, "CErc20 mint failed.");
+            uint256 cTargetTokenMintResult = cTargetToken.mint(
+                postActionData.toAmount
+            );
+            require(
+                cTargetTokenMintResult == SUCCESS_CODE,
+                "CErc20 mint failed."
+            );
 
-            uint postActionFinalBalance = cTargetToken.balanceOf(address(this));
+            uint256 postActionFinalBalance = cTargetToken.balanceOf(
+                address(this)
+            );
 
-            uint cAssetTransferredBalance = calculateAndTransferCErc20To(postActionData, cTargetToken, postActionInitialBalance, postActionFinalBalance);
+            uint256 cAssetTransferredBalance = calculateAndTransferCErc20To(
+                postActionData,
+                cTargetToken,
+                postActionInitialBalance,
+                postActionFinalBalance
+            );
 
-            emitActionExecutedEvent(postActionData, cTargetTokenAddress, cAssetTransferredBalance);
+            emitActionExecutedEvent(
+                postActionData,
+                cTargetTokenAddress,
+                cAssetTransferredBalance
+            );
         } else {
             revert("Target token is not supported in Compound.finance.");
         }
@@ -99,15 +130,17 @@ contract CompoundMintPostAction is PostActionBase {
     function calculateAndTransferCErc20To(
         StablePayCommon.PostActionData memory postActionData,
         CErc20 cTargetToken,
-        uint postActionInitialBalance,
-        uint postActionFinalBalance
-    )
-        internal
-        returns (uint)
-    {
-        uint postActionCAssetBalance = postActionFinalBalance.sub(postActionInitialBalance);
+        uint256 postActionInitialBalance,
+        uint256 postActionFinalBalance
+    ) internal returns (uint256) {
+        uint256 postActionCAssetBalance = postActionFinalBalance.sub(
+            postActionInitialBalance
+        );
         require(
-            cTargetToken.transfer(postActionData.toAddress, postActionCAssetBalance),
+            cTargetToken.transfer(
+                postActionData.toAddress,
+                postActionCAssetBalance
+            ),
             "Transfer to 'to' address failed."
         );
         return postActionCAssetBalance;
@@ -119,9 +152,11 @@ contract CompoundMintPostAction is PostActionBase {
         @param cTargetTokenAddress target token address used in Compound.finance platform.
         @param cAssetTransferredBalance target amount minted in Compound.finance.
      */
-    function emitActionExecutedEvent(StablePayCommon.PostActionData memory postActionData, address cTargetTokenAddress, uint cAssetTransferredBalance)
-        internal
-    {
+    function emitActionExecutedEvent(
+        StablePayCommon.PostActionData memory postActionData,
+        address cTargetTokenAddress,
+        uint256 cAssetTransferredBalance
+    ) internal {
         emit ActionExecuted(
             address(this),
             postActionData.sourceAmount,
