+++ contracts/services/compound/CErc20.sol
@@ -8,7 +8,6 @@ import "./CToken.sol";
  * @author Compound
  */
 contract CErc20 is CToken {
-
     /*** User Interface ***/
 
     /**
@@ -17,7 +16,7 @@ contract CErc20 is CToken {
      * @param mintAmount The amount of the underlying asset to supply
      * @return uint 0=success, otherwise a failure (see ErrorReporter.sol for details)
      */
-    function mint(uint mintAmount) external returns (uint);
+    function mint(uint256 mintAmount) external returns (uint256);
 
     /**
      * @notice Sender redeems cTokens in exchange for the underlying asset
@@ -25,7 +24,7 @@ contract CErc20 is CToken {
      * @param redeemTokens The number of cTokens to redeem into underlying
      * @return uint 0=success, otherwise a failure (see ErrorReporter.sol for details)
      */
-    function redeem(uint redeemTokens) external returns (uint);
+    function redeem(uint256 redeemTokens) external returns (uint256);
 
     /**
      * @notice Sender redeems cTokens in exchange for a specified amount of underlying asset
@@ -33,21 +32,21 @@ contract CErc20 is CToken {
      * @param redeemAmount The amount of underlying to redeem
      * @return uint 0=success, otherwise a failure (see ErrorReporter.sol for details)
      */
-    function redeemUnderlying(uint redeemAmount) external returns (uint);
+    function redeemUnderlying(uint256 redeemAmount) external returns (uint256);
 
     /**
       * @notice Sender borrows assets from the protocol to their own address
       * @param borrowAmount The amount of the underlying asset to borrow
       * @return uint 0=success, otherwise a failure (see ErrorReporter.sol for details)
       */
-    function borrow(uint borrowAmount) external returns (uint);
+    function borrow(uint256 borrowAmount) external returns (uint256);
 
     /**
      * @notice Sender repays their own borrow
      * @param repayAmount The amount to repay
      * @return uint 0=success, otherwise a failure (see ErrorReporter.sol for details)
      */
-    function repayBorrow(uint repayAmount) external returns (uint);
+    function repayBorrow(uint256 repayAmount) external returns (uint256);
 
     /**
      * @notice Sender repays a borrow belonging to borrower
@@ -55,7 +54,9 @@ contract CErc20 is CToken {
      * @param repayAmount The amount to repay
      * @return uint 0=success, otherwise a failure (see ErrorReporter.sol for details)
      */
-    function repayBorrowBehalf(address borrower, uint repayAmount) external returns (uint);
+    function repayBorrowBehalf(address borrower, uint256 repayAmount)
+        external
+        returns (uint256);
 
     /**
      * @notice The sender liquidates the borrowers collateral.
@@ -65,5 +66,9 @@ contract CErc20 is CToken {
      * @param repayAmount The amount of the underlying borrowed asset to repay
      * @return uint 0=success, otherwise a failure (see ErrorReporter.sol for details)
      */
-    function liquidateBorrow(address borrower, uint repayAmount, CToken cTokenCollateral) external returns (uint);
+    function liquidateBorrow(
+        address borrower,
+        uint256 repayAmount,
+        CToken cTokenCollateral
+    ) external returns (uint256);
 }
